+++ inc/projects.php
@@ -1,5 +1,5 @@
 <?php
-
+session_start();
 function amsprojects_function( $slug ) {
     ob_start();  
     ?>
@@ -46,15 +46,12 @@ else
    $blockclass = 'main-content-three-col';
 }
 
+
 ?>
 
 <div class="wp-block-columns main-content <?= $blockclass; ?>" >
 
   <?php
-
-  /*echo "<pre>";
-  print_r($gridlayout);
-  echo "</pre>";*/
 
   if (!isset($blockdata['projectsidebar']))
   {        
@@ -91,7 +88,7 @@ else
           
           $synopsis = mb_strimwidth($x_value['synopsis'], 0, 150, '...');
             
-          echo "<div class='listview-project'>";
+          echo "<div class='listview-project projectdiv'>";
           echo "<div class='assets-list-items'>";
           
 
@@ -113,14 +110,28 @@ else
                 }
 
           
-          echo "<div class='assetsproduct-content'><a href='".site_url('/project/'.$x_value['id'].'-'.$x_value['user_id'].'-'.$pageid)."'>";
-          echo  "<p class='product-title'> ". $x_value['name'] ;
-          if($x_value['completed_year'])
+          echo "<div class='assetsproduct-content'>";
+          if($_SESSION["projectpassword"] || $blockdata['project_protected'] == NULL)
           {
-            echo " (".$x_value['completed_year'].")";
+            echo "<a href='".site_url('/project/'.$x_value['id'].'-'.$x_value['user_id'].'-'.$pageid)."'>";
+            echo  "<p class='product-title'> ". $x_value['name'] ;
+            if($x_value['completed_year'])
+            {
+              echo " (".$x_value['completed_year'].")";
+            }
+            echo "</p>";
+            echo  "</a>";
           }
-          echo "</p>";
-          echo  "</a>";
+          else
+          {
+            echo  "<p class='product-title'> ". $x_value['name'] ;
+            if($x_value['completed_year'])
+            {
+              echo " (".$x_value['completed_year'].")";
+            }
+            echo "</p>";
+          }  
+
           echo "<div class='assetsprice'>";
           echo    "<p class='memberprice'><strong>Created By</strong> - ". $x_value['creator']. "</p>";
 
@@ -149,9 +160,29 @@ else
         foreach($arrayResult['projects'] as $x_value) 
         {
           echo"<div class='productstyle projectdiv'>";
-                echo "<a href='".site_url('/project/'.$x_value['id'].'-'.$x_value['user_id'].'-'.$pageid)."'>";
-                echo  "<p class='product-title'>".$x_value['name']. " (2005)</p>";
-                echo "</a>";
+               
+                if($_SESSION["projectpassword"] || $blockdata['project_protected'] == NULL)
+                {
+                  echo "<a href='".site_url('/project/'.$x_value['id'].'-'.$x_value['user_id'].'-'.$pageid)."'>";
+                  echo  "<p class='product-title'>".$x_value['name'];
+                  if($x_value['completed_year'])
+                  {
+                    echo " (".$x_value['completed_year'].")";
+                  }
+                  echo "</p>";
+                  echo "</a>";
+                }
+                else
+                {   
+                  echo  "<p class='product-title'>".$x_value['name'];
+                  if($x_value['completed_year'])
+                  {
+                    echo " (".$x_value['completed_year'].")";
+                  }
+                  echo "</p>";
+                  
+                } 
+
                 if($x_value['thumbnail'] == NULL || $x_value['thumbnail'] == "")
                 {                                    
                     echo "<div class='product-img-wrap'>";
@@ -193,7 +224,24 @@ else
    
 </div>
 
-
+<!-- Popup -->
+<div class="custom-model-main">
+    <div class="custom-model-inner">        
+    <div class="close-btn">×</div>
+        <div class="custom-model-wrap">
+            <span id="amscredentials_error"></span>
+            <div class="pop-up-content-wrap">
+               <input type="text" name="projectpassword" id="projectpassword">
+               <input type="submit" id="projectsubmit">
+                <span class="customprojectloader" id="projectinifiniteLoader">
+                  <img src="<?php echo esc_url( plugins_url( 'assets/img/buttonloader.gif', dirname(__FILE__) ) ) ?>">
+                </span>
+            </div>
+        </div>  
+    </div>  
+    <div class="bg-overlay"></div>
+</div>
+<!-- End popup -->
 
 <script type="text/javascript">
 jQuery(document).ready(function($) {
@@ -203,6 +251,8 @@ jQuery(document).ready(function($) {
    var pageid = jQuery("#getpageid").val();
 
    $('#inifiniteLoader').hide();
+   jQuery(".customprojectloader").hide();
+   amsblocklogin();
 
     $('#seemore').click(function(){
 
@@ -254,13 +304,73 @@ jQuery(document).ready(function($) {
             $("#inifiniteLoader").show();
        },
        success: function (html) {
-         jQuery('#inifiniteLoader').hide('1000');
+         jQuery('#inifiniteLoader').hide();
          jQuery('.right-col-wrap').append(html);
          jQuery('#seemore').show();
+         amsblocklogin();
        }
      });
      return false;
     }
+
+    // Custom popup login
+    function amsblocklogin()
+    {
+      var projectpasswordsession = "<?php echo $_SESSION["projectpassword"]; ?>"; 
+      var project_protected = "<?php echo $blockdata['project_protected']; ?>";
+      if(projectpasswordsession == '' && project_protected != '')
+      {
+        jQuery(".projectdiv").on('click', function() {
+          jQuery(".custom-model-main").addClass('model-open');
+        }); 
+        jQuery(".close-btn, .bg-overlay").click(function(){
+          jQuery(".custom-model-main").removeClass('model-open');
+        });
+      } 
+    }
+
+    $('#projectsubmit').click(function(){
+        
+        var projectpassword = jQuery('#projectpassword').val();
+        var getpageid = jQuery('#getpageid').val();
+        
+          $.ajax({
+             url: amsjs_ajax_url.ajaxurl,
+             type:'POST',
+             data: { action: 'get_amsprojectlog', projectpassword:projectpassword,getpageid:getpageid},
+             beforeSend: function(){
+              // Show image container
+                jQuery(".customprojectloader").show();
+                jQuery("#projectsubmit").prop('disabled', true);
+             },
+             success: function (data) {
+               /* console.log(data);*/
+                jQuery('.customprojectloader').hide('1000');
+                jQuery("#projectsubmit").prop('disabled', false);
+
+                var mydata = data.substring(0,data.length - 1);
+
+                console.log(mydata);
+
+                 if(mydata == 'valid')
+                 {
+                    location.reload();
+                 }
+                 else
+                 {
+                    jQuery("#amscredentials_error").html('<p>AMS Credentials not match.</p>');
+                    jQuery("#amscredentials_error").css("color", "red");
+                    jQuery("#amscredentials_error").css("display", "block");
+
+                    setTimeout(function() {
+                        $('#amscredentials_error').fadeOut('fast');
+                    }, 5000);
+                    
+                 }
+             }
+          });
+    
+    });
    
 });    
 </script>
